requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_pytorch_select=0.1=cpu_0
aiohttp=3.7.4=py37h27cfd23_1
async-timeout=3.0.1=py37h06a4308_0
attrs=20.3.0=pyhd3eb1b0_0
blas=1.0=mkl
boto3=1.17.46=pyhd3eb1b0_0
botocore=1.20.50=pyhd3eb1b0_1
brotlipy=0.7.0=py37h27cfd23_1003
ca-certificates=2021.4.13=h06a4308_1
cachetools=4.2.1=pyhd3eb1b0_0
cairo=1.14.12=h8948797_3
certifi=2020.12.5=py37h06a4308_0
cffi=1.14.0=py37h2e261b9_0
chardet=3.0.4=py37h06a4308_1003
click=7.1.2=pyhd3eb1b0_0
cryptography=3.4.7=py37hd23ed53_0
cycler=0.10.0=py37_0
dbus=1.13.18=hb2f20db_0
decorator=4.4.2=pyhd3eb1b0_0
expat=2.3.0=h2531618_2
fontconfig=2.13.1=h6c09931_0
freetype=2.10.4=h5ab3b9f_0
fribidi=1.0.10=h7b6447c_0
gensim=3.8.3=py37h2531618_2
glib=2.63.1=h5a9c865_0
google-api-core=1.25.1=pyhd3eb1b0_0
google-auth=1.29.0=pyhd3eb1b0_0
google-cloud-core=1.6.0=pyhd3eb1b0_0
google-cloud-storage=1.37.1=pyhd3eb1b0_0
google-crc32c=1.1.2=py37h27cfd23_0
google-resumable-media=1.2.0=pyhd3eb1b0_1
googleapis-common-protos=1.53.0=py37h06a4308_0
graphite2=1.3.14=h23475e2_0
graphviz=2.40.1=h21bd128_2
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb453b48_1
harfbuzz=1.8.8=hffaf4a1_0
icu=58.2=he6710b0_3
idna=2.10=pyhd3eb1b0_0
intel-openmp=2019.4=243
jmespath=0.10.0=py_0
joblib=1.0.1=pyhd3eb1b0_0
jpeg=9b=h024ee3a_2
kiwisolver=1.3.1=py37h2531618_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.33.1=h53a641e_7
libcrc32c=1.1.1=he6710b0_2
libffi=3.2.1=hf484d3e_1007
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libmklml=2019.0.5=0
libpng=1.6.37=hbc83047_0
libprotobuf=3.14.0=h8c45485_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
libuuid=1.0.3=h1bed415_2
libxcb=1.14=h7b6447c_0
libxml2=2.9.10=hb55368b_3
lz4-c=1.9.3=h2531618_0
matplotlib=3.3.4=py37h06a4308_0
matplotlib-base=3.3.4=py37h62a2d02_0
mkl=2020.2=256
mkl-service=2.3.0=py37he8ac12f_0
mkl_fft=1.3.0=py37h54f3939_0
mkl_random=1.1.1=py37h0573a6f_0
multidict=5.1.0=py37h27cfd23_2
ncurses=6.2=he6710b0_1
networkx=2.5.1=pyhd3eb1b0_0
ninja=1.10.2=hff7bd54_1
nltk=3.6.1=pyhd3eb1b0_0
numpy=1.19.2=py37h54aff64_0
numpy-base=1.19.2=py37hfa32c7d_0
olefile=0.46=py37_0
openssl=1.1.1k=h27cfd23_0
pandas=1.2.4=py37h2531618_0
pango=1.42.4=h049681c_0
pcre=8.44=he6710b0_0
pillow=8.2.0=py37he98fc37_0
pip=21.0.1=py37h06a4308_0
pixman=0.40.0=h7b6447c_0
protobuf=3.14.0=py37h2531618_1
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.8=py_0
pycparser=2.20=py_2
pydot=1.4.1=py37h06a4308_0
pyopenssl=20.0.1=pyhd3eb1b0_1
pyparsing=2.4.7=pyhd3eb1b0_0
pyqt=5.9.2=py37h05f1152_2
pysocks=1.7.1=py37_1
python=3.7.7=hcf32534_0_cpython
python-dateutil=2.8.1=pyhd3eb1b0_0
python-graphviz=0.16=pypi_0
pytorch=1.7.1=cpu_py37h6a09485_0
pytz=2021.1=pyhd3eb1b0_0
qt=5.9.7=h5867ecd_1
readline=8.1=h27cfd23_0
regex=2021.4.4=py37h27cfd23_0
requests=2.25.1=pyhd3eb1b0_0
rsa=4.7.2=pyhd3eb1b0_1
s3transfer=0.3.6=pyhd3eb1b0_0
scikit-learn=0.24.1=py37ha9443f7_0
scipy=1.6.2=py37h91f5cce_0
setuptools=52.0.0=py37h06a4308_0
sip=4.19.8=py37hf484d3e_0
six=1.15.0=py37h06a4308_0
smart_open=5.0.0=pyhd3eb1b0_0
sqlite=3.35.4=hdfb4753_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.10=hbc83047_0
tornado=6.1=py37h27cfd23_0
tqdm=4.59.0=pyhd3eb1b0_1
typing-extensions=3.7.4.3=hd3eb1b0_0
typing_extensions=3.7.4.3=pyh06a4308_0
urllib3=1.26.4=pyhd3eb1b0_0
wheel=0.36.2=pyhd3eb1b0_0
xz=5.2.5=h7b6447c_0
yarl=1.5.1=py37h7b6447c_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.9=haebb681_0
